Fix up bigram generation
Bigram generation needs to be done on the Perseus wordlist before
it has been processed, which we weren't doing previously.

Also set a threshold on the bigrams so that the training doesn't
get too big, and only the most common ~150,000 are included.

Also cleaned up the bigram script a little.