Don't bother with betacode words from wordlist
The betacode words took some work to convert to UTF-8, and were
hard to get right across multiple platforms. They only account for
about 1% of the Greek we're getting from Perseus, so best is just
to drop them and enjoy the extra simplicity.

Also I discovered that character classes in regular expressions are
locale dependent, so set LC_ALL=C is set for an awk script that
uses them.