Rewrite makegarbage
It's much more efficient and faster now, and more sensible,
but most importantly it is much clearer code.

It also removes the dependency on isupper, which removes the
dependency on runetype, which is all good.

Also change the LC_ALL=C back for case of wordlist.perseus, as
it turns out the exact meaning of character ranges in regexp are
officially 'undefined' as of posix 2008, and the LC_ALL=C made
output not UTF-8. It would be nice to have the script be locale
agnostic, but I don't think that's possible with awk, but it should
work anywhere the locale is UTF-8 based, which is anywhere that is
sane, which is good enough.